+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - return content and `Config\Loader\FileLoader` does not store it
   - merge content of all included files and remove imports
 - `Util\Math::getTwoTablePagination()` is static now
+- `Config\Util\ArrayHelper` moved to `Util\ArrayHelper`
 - [git diff](/../../compare/1.2...master)
 
 ### Added